Remove the shutdown grace period
This removes the `vast.shutdown-grace-period` functionality and its underlying
mechanism. This has two reasons:

1. The mechanism was broken, and in case of yet unpersisted data taking longer
   than the grace period to persist on shutdown caused both data loss and an
   infinite hang.

2. It is the purpose of a service manager to issue hard kills to hanging
   process, not of the hanging process itself.

Ultimately we decided that the internal hard kill causes more issues than it
helps solves, so now the new grace period is infinite: VAST simply waits for as
long as it takes to shut down properly.