Re-enable plugin tests in macOS CI
We recently disabled testing for the plugins in the macOS CI due to test
flakiness; that flakiness, however, we were not able to reliably reproduce after
a few days. This re-enables the building of plugins in the macOS CI.

The plan is to run the CI multiple times before merging this, just so we can be
somewhat sure that the issue no longer exists.