Fix shutdown hang in the filesystem actor
This fixes a regression I introduced myself last week when adding metrics to the
filesystem actor: A cycle between the accountant and the filesystem that caused
an infinite hang on shutdown when metrics are enabled. We didn't catch this in
integration tests because those, well, run with metrics disabled.

Because I didn't want to add a `actor_cast<T>(some_weak_actor_ptr)` to our code
base in a seemingly random place, I added a typed abstraction for weak actor
handles to our code base, which feels like a thing that we should contribute
upstream as `typed_actor<...>::weak_handle` or similar.