Provide an Ansible role for VAST
This adds a relatively simple role for deploying VAST to a Linux
host with Ansible. It supports FHS-based distributions in general,
but Debian and derivatives are tested the most.

This role has been in use at Tenzir for a while and we hope to
provide additional value to our users by adding it the the open
source distribution.