Make it less likely for rebuild to require rebatching
Scenario: We have the partitions A and B, and automatic rebuilding merges them
into the partition AB. All except for the last batch in AB are perfectly sized.

Now comes partition C with undersized batches. If the rebuilder operates in the
order C -> AB it is more likely to have to rebatch than if it were to operate in
the order AB -> C, because the beginning of AB is already guaranteed not to
require rebatching.

This change makes it so we always rebuild the selected partitions for a given
rebuild run in time ascending order, while keeping the selection order for
rebuilding to be time descending.