Make enum extension types work
Turns out that pa.array() with type=T where T is a dictionary type is not
sufficient to use the proper type for the dictionary indices. We had uint8 in
our type, but unless we explicitly go through pa.DictionaryArray.from_arrays, we
will get a int8. This type clash was the issue in the past.