Fix version passing for the static binary
The recent version fallback refactoring exposed an issue in the
Nix packaging the made it possible to prefix the version with two
'v's. We need to make sure to strip the leading 'v' for the Nix
package version, then we can re-prepend it when passing the number
into the build scaffold.