Enable use of env variables for web plugin options
Before this change, it was impossible to set options for the web plugin using
environment variables, and in the configuration file the category they had to be
specified in the wrong category.

This change fixes that, enabling use of `VAST_PLUGINS__WEB__MODE=dev vast web
server` and similar uses, which comes in handy for everything Docker-related.