Fix race in plugin tests for CTest with -j2 or higher
I just noticed locally when running `ctest --test-dir build -j8` that there was
a race for building libvast, which happens because we have test fixtures for
vast-test and the individual plugin unit test binaries that ensure that we build
the plugins. However, there was a race between the fixture setup the individual
plugin unit test binaries in case libvast was not yet built. This change fixes
that by making the dependency more explicit, and ensuring that the fixture for
the vast-test binary has priority when running CTest.