Integrate Cortex Analyzer for VAST [WIP]
This almost works, however the container that runs the actual query
(tenzir/vast-neuron) doesn't find the `input.json` in the expected place
(`/job/input/input.json`). I checked that the `input.json` is correctly
creatd in the cortex container with proper contents, and when the
container is started an actual log message indicates that that very
volume is mounted to `/job` on the neuron.

```
thehive-cortex-1         |   timeout: 30 minutes
thehive-cortex-1         |   image  : tenzir/vast-neuron
thehive-cortex-1         |   volume : /opt/cortex/jobs/cortex-job-yfueFYQBc4_qCE3nWixQ-7660690682055716688:/job
```

However, when logging into the neuron /job is empty. I've run out of
ideas. It does seem to work for Kaans set of images, maybe some sort of
permission problem.