Send 'atom::done' to the sink instead of the sender
The index responds to a query by sending a number of
table slices followed by an `atom::done`. However, while
the slices are sent to the sink specified in the query context,
the final 'done' was sent to the sender of the original
query request.

For existing usages these were always identical, except for the
partition transformer which is implemented in the index itself
and thus has access to the internals and changes the target
of the 'done' atom manually.