Rename `vast dump` to `vast show`
We're planning to do more with the mostly internal command in the near
future, and to quote @lava on the matter:

> I wanna take the opportunity to suggest renaming `vast dump` to `vast
> show`, since `dump` sounds like some kind of backup or debug command.

I very much agree with that sentiment. This change implements the
transition without any deprecation period as the command is not even
documented currently and I do not anticipate anyone actually depending
on it.