Allow read access to user home dir in the systemd unit
In case the XDG config directory of the vast user is not in
`/var/lib/vast` the sandboxing configuration of the unit would
deny access to the configuration files in that place, causing
the main process to exit.

In theory the config directory may be outside of the home directory,
but systemd has no mechanism to specify that, so we make the home
dir read only instead.