Simplify artifact upload name for static builds
Now that we include the epoch number the package name contains a colon, which is
not an allowed character in the name for the upload-artifact action. Since this
name is only used in the GitHub Actions UI or when dependent jobs download the
artifact (which they do not) we can just change it to whatever we want.