Implement pipelines according to RFC-001
A pipeline is an ordered sequence of logical operators with matching
input and output types along the sequence. The pipeline's input type is
the input type of the first operator in the sequence, and the pipeline's
output type is the output type of the last operator in the sequence. A
pipeline is by definition a logical operator Instantiating a pipeline
directly is forbidden.