Infer non-config types in extend and replace operators
Previously, using an address value in the extend or replace operators
did not work. This is now fixed, and will produce the proper type as
long as it can be inferred from the data. This bug applied to all types
that were not also used as types in configuration, i.e., address,
subnet, pattern, and map.