Remove the `#field` meta extractor
The predicate `#field == "X"` matched all events for which the field
extractor `X` had data, i.e., it was semantically equivalent to `X !=
nil`.

Fundamentally, the `#field` meta extractor broke our model of
extractors, since it did not result in a single value or a list of
values to compare against, but rather create an implicit disjunction
with changed relational operator semantics compared to all other
predicates.

This removes the operator without a deprecation period as it received
very little usage.