Remove outdated table slice encodings
Following the introduction of a minimum partition version of 1 with
tenzir/vast#2778, we're no longer able to encounter any table slices
that are MessagePack-encoded or use an older version of the mapping of
our data model onto Apache Arrow.

There's a future refactoring that is now unblocked, which is a total
replacement of the legacy baggage that is the `table_slice` API with an
easier-to-use wrapper around the same underlying data that makes more
assumptions about the underlying column-major memory layout.

This is based on tenzir/vast#2797 to avoid merge conflicts.