Make feather stores read incrementally
This changes `feather` stores not to use the official Apache Feather
reader anymore, but rather use its lower-level building blocks to do
reads incrementally. This means that loading a feather store should not
take any time anymore; the actual reading only happens on demand as
batches are accessed. This is implemented in a way that the store caches
already read batches so they're not read again for concurrent queries.

NB: This technically drops support for reading Apache Feather V1 files,
but VAST never wrote them to begin with and we never officially said we
support that.