Fix UB in `order(...)` for floating point numbers
The `coder` unit tests using the `order(...)` function on floating point
numbers started failing for release builds on Debian after the CAF 0.18
bump, but the unit test wasn't touched since 2016, and the code itself
not since 2018, so it was pretty like to be undefined behavior that we
ran into. That turned out be rather easy to spot looking at the
function.urned out be rather easy to spot looking at the function.