Avoid calling the chunk deleter twice
We're seeing some weird ASan failures in CI that we cannot really explain or
reliably reproduce, but they all look like they could be caused by a chunk's
deleter being invoked twice. This can only happen if the destructor is called
twice, which doesn't make a lot of to me, but let's avoid that being an issue by
employing defensive programming techniques and observing whether the issue
occurs again.