Force a flush at the end of blocking imports
With the faster termination thanks to the new signal handling logic
the symptoms of the not quite working blocking import become much
more pronounced. This is especially problematic for the integration
tests which rely on blocking import a lot.

In order to preserve the blocking semantic the flag now implies a
flush to disk of in memory data. This ensures that active partitions
are written to disk and added catalog before the command returns.