Change boolean literals to `true` and `false`
Historically, VAST used the literals `T` and `F` to denote true and
false values. This change switches to `true` and `false` instead, which
are much more widely used—essentially everywhere.

I asked for feedback on this in the community Slack and this appeared to
be a much appreciated change. Given that the next release is a major
release this is without a deprecation period. It also changes the
printing dual of booleans in VAST's native export format "ascii" to
`true` and `false` accordingly.