Fix some shutdown issues in the server command
This commit moves the web server into a separate thread
so that the main thread can wait for DOWN messages and
signal notifications in the meantime.

It also fixes an issue where the spawned handlers would
ignore an `exit_msg` with a default-constructed caf::error;
we now send a user_shutdown message instead.