Fix bug caused Arrow imports to skip buffered events
This is a two-fold commit: First, the Arrow reader threw away stream readers
when the number of requested events were exceeded. In case the reader had more
slices buffered this led to data being thrown away. This did not completely fix
the race in the integration tests, because there appears to be a race condition
when running the Arrow import with a local server process by setting the
`vast.node` option.

The bug became much rarer after the first change, so in order to make the tests
more reliable I also changed the tests not to use a local server process.
Afterwards, I was unable to reproduce the issue in over a thousand repetitions
of the Arrow Import test in release mode, when without any of the two changes it
occurred every ~10th time and with the first fix every ~50th time on my local
machine with a release build.