Remove the `pattern` type
This follows up on the removal of the match operator and introduction of
pattern search using predicates of the form `lhs == /rhs/` by removing
the pattern type without replacement.

In theory this is a hard and incompatible breaking change. However, we
know of no single user that has relied upon the pattern type being
available, and as such it _should_ be safe to be removed without a
transparent upgrade path.

To make clear just how unlikely it is that anyone ran into this: users
would've had to have a field of type `pattern #index=skip` in their
schema, as attempting to index a pattern field crashed because we never
actually implemented support for it, which was (1) never documented and
(2) nobody ever complained about it.