Introduce an extensible mechanism for casting data
This introduces a new set of functions `vast::cast` and
`vast::can_cast`, with the former operating on arrays and the latter
operating on types. For now, this only supports reshuffling columns in
record types (and corresponding struct arrays), but the mechanism is
written in a way that is easily extensible via template specializations.