Align output of the Zeek TSV reader with schemas
This changes the Zeek TSV reader to respect available schemas from
schema files in case the parsed output can be casted to a schema with an
exactly matching name.

In case the casting is impossible, the Zeek reader warns exactly once
when switching to a new file and ignores the user-provided schema.