Re-order Zeek schemas
This removes the `_path` from the Zeek schemas as they are only required for
disambiguation in the Zeek JSON reader, but should not actually be part of the
data. It also moves the `_write_ts` field to the end to increase compatibility
with downstream tooling that relies on the column order in Zeek tools (please
use `zeek-cut`, folks).