Pass table slices through pipeline operators
This lowers the boundary to writing additional operators, as falling
down to the lower-level Arrow APIs is now opt-in rather than being
required. It also makes it so that pipeline operators that do not create
new data (e.g., summarize) can keep the import timestamps unchanged,
which fixes a long-standing issue that caused compaction and rebuilding
to shift import timestamps.

Most importantly, this will make the transition to the upcoming
rewritten executor and its pipeline operator API changes a lot easier.