Fix a crash when running import pipelines
This was really a headscratcher: The constructor of table slice used
when setting the import time in a slice that's not unqiue (1) was
unnecessarily expensive and (2) only worked if the slice was serialized
and not changed since.

Uisng server-side import pipelines actually resulted in a crash before
this change: the pipelines resulted in slices that did not already have
a serialized backing, which is a precondition for using the constructor
that creates a slice from such a serialized backing.