Improve python bindings ergonomics
This changes the API to go through an object of the `VAST` type
instead of working directly on the type. This way it is possible
to specify certain option parameters like the connection endpoint
for a group of commands without requiring filesystem or process
level state.

This makes it also possible to run the integration tests of the
bindings without docker.