Add a workaround to fix CAF OpenSSL options
In the `actor_system_config`, some settings are not like the others.
While most modules read their options directly from the `content`
object, the SSL configuration is stored in member variables of the
main config class. The CAF command line and config file parsers
handle this correctly, but settings from `vast.yaml` files or those
that are passed in as environment variables are injected through
a different code path.

This commit adds assignments to set the aforementioned class members
explicitly.

I considered the alternative solution of translating non CAF-native
option formats to native ones and let the `parse` function handle
it, but that approach would be a large amount of work rather little
benefit.