Support pipeline syntax in compaction
Compaction uses this one function that's not used anywhere else to build
pipelines, and we forgot that in the earlier PR that added support for
the new syntax for pipelines defined in the `vast.pipelines` section in
the config file. This fixes that oversight.

Ultimately, we should just get rid of this function, but let's fix what
we got for now and then rewrite it when we get back to this.