Remove the frontend prototype bundled with the web plugin
This removes the existing frontend prototype from the web plugin that
was bundled with the plugin installation and served as the default
document root. Some parts of the frontend that we have in development
are designed to be closed-source, and it is easier to develop at the
current development stage in a single repository that is not bound to
the release process of VAST itself.

An open-source version of the frontend may return in the future.