Introduce the operator control plane
This is still early in design, but we have essentially agreed upon the
requirements for the operator control plane, i.e., a set of escape
hatches for the individual operators. This is purely virtual so it can
be easily mocked in unit tests for individual operators. In the RFC-001
implementation of the pipeline executor, it is the executor's
responsibility to provide an instance of this control plane to every
operator.