Add the --taste-size option to vast export
The new option allows the user to set the maximum number of
partitions that are part of the initial evaluation batch.
This functionality has been present internally already, but was
not exposed on the command line until now.

This is particularly useful for queries that contain pipeline-breaking
operators, as those can yield unexpected results with the default
taste size.