Implement the `vast exec` command
This command takes a single positional argument that must parse as a
valid logical pipeline, and then immediately executes it.

The current implementation is pretty naïve, but it suffices for initial
testing once we have the from and to operators in place to actually be
able to write a closed pipeline.

I've written a follow-up issue to implement signal handling in the
executor; I've intentionally not implemented that for now as the API for
the exeuctor will change to an asynchronous rather than just concurrent
API soon after this initial version, which makes it easier to implement
signal handling correctly.

Fixes tenzir/issues#202