Log handled failures to persist partition synopses
When a partition synopsis is not present, VAST warns about it on startup
attempts to restore it. However, we did not log when serializing or
persisting that synopsis failed, which is the information you're looking
for first when something related went wrong. This adds such logs,
clearly indicating that the data will be restored on the next start of
the VAST node.