Fix log message listing candidate partitions
The catalog has for a while now shown a log message showing the number
of candidate partitions to be considered for an incoming query. A few
releases back we changed the catalog to do the lookup per schema, and
then return the aggregated result. The log message was then shown per
schema, rather than just once.

There are two ways to fix this:
1. Continue showing the log message once per schema, but actually show
   which schema it is for.
2. Show the log message once in total.

I chose to go for (2) as that aligns the log message with the
corresponding `catalog.lookup.candidates` metric.