Introduce the `put` operator
The `put` operator is a replacement for both the `extend` and the
`replace` operators. It takes as arguments a comma-separated list of
assignments. If the assigned-to extractor resolves for the input data
replaces all occurences of the field (previously `replace`), and if the
extractor does not resolve adds one new field at the end (previously
`extend`).

`put` is the first operator not to be available in the
old pipeline operator framing, and as such is not currently reachable
by the user.