Warn when TheHive integration tests fail
Currently our CI fails a lot because of a known transient error with
TheHive. We ignore this error code, but still the CI shows as failed a
lot of the time, making the indicator rather worthless. This changes the
setup so the tests doesn't actually fail in this scenario, but rather
emits a warning.