Fix partition selection for the `rebuild` command
This fixes the `rebuild` command not to loop infinitely over all
partitions when at least one not-emprty set of partitions needed to be
rebuilt.

This logic bug was introduced with the VAST v3.0.2 release, and can be
tracked down to the confusing meaning of the `--all` flag. In
addition, this also changes the meaning of the option to do what it
should, namely rebuild _all_ partitions instead of just _outdated_ ones.
The `--undersized` flag remains unchanged and still adds all undersized
partitions to the list of partitions considered for rebuilding,
regardless of whether they're outdated or not.