Implement a `cast` operator
The semantics of this are rather straightforward: The `cast <schema>`
operator takes the name of a known schema, and parses input events to it
iff possible.

I developed this primarily because I needed it to be able to test
casting to understand how we want to improve upon the adaptive builder,
and figured that it'd make for a pretty good operator on its own.