Implement a script for timeshifting zeek logs
Takes a path to a Zeek log and shifts the timestamp values so that the
first event is now.

Caveats:
- Assumes a fixed header length of 8 lines
- Does not rewrite the header timestamp
- is not really accurate, because the processing time of the code itself
  is not taken into account
- lacks an option to not delay the output
- lacks an option to set a fixed starting time

This should ideally be a VAST operator, but until we have that this
should be somewhat usable. It can also serve as a test case for
implementing the operator if we remove the last two caveats.