Add a `--mmap` option to the `file` loader
This option is for experts: it allows for memory-mapping a file rather
than reading it into memory incrementally. This is a trade-off: Instead
of parallelizing reads, this enables the parser to control reads without
knowing that it does.

For the Feather and Parquet parsers, this reduces the memory usage and
improves the performance significantly.