Ensure reliable stream shutdown with in-stream sentinel values
After a lot of experimenting and researching throughout the CAF code
base, it turned out that the only reliable way to shut down a stream
spanning actors in multiple processes is to send a sentinel value within
the stream itself.

Now, you may wonder, why don't the other approaches work? Here's why:

1. Make the stream own the execution nodes: This works wonderfully up to
   the point where an execution node is in a different process, and its
   reference count never goes to zero when the part of the stream hosted
   by that execution node goes down.

2. Close and flush the stream explicitly: This causes a race condition
   in the stream drivers finalize handler, which makes it impossible to
   send out further events from a stream stage to a stream sink after
   the stream has been closed.

3. Shut down the hosting execution node actor from within the stream:
   This causes another race condition where it the stream finalize
   handler may be called after the actor state was destroyed, causing
   address sanitizer to go boom.

So yeah, in-stream sentinel values it is. Works wonderfully.