Implement the `import` operator
The `import` operator is a remote event sink for a pipeline that imports
events into VAST.

This current implementation is very basic, and simply forwards events to
the existing `importer` singleton component in the VAST node.

In the future, we should change the `import` operator to directly create
passive partitions, and to send then to the catalog directly, which will
enable better fault tolerance. That, however, is a change we can make in
the future, as VAST's internal APIs are not well suited for that at the
current point in time.

To make testing easier, this also implements a `repeat` operator that
repeats an input (either bytes or events) a given number of times.