Fix invalid access to operator in an execution node
This simplifies the stream drivers for the execution node to no longer
receive ownership of the operators and to instead read them from the
execution node state. The old mechanism could in some scenarios lead to
a crash when calling `ctrl.warn(...)`, for which the precondition
`state.op_ != nullptr` was sometimes violated.