Preserve nesting structure of pipelines
This commit makes three changes to the internals of pipelines, and has
an additional nicety for local pipeline execution:

1. We no longer flatten pipelines used as operators at construction time
   of a pipeline.
2. Pipelines now preserve the definition they were created from.
3. Unwrapping a pipeline no longer requires destroying the pipeline, and
   now returns an offset into the nested pipeline structure as well as a
   pointer to the unwrapped operator.

In the future, these changes will allow us to track how operators are
shuffled around as part of pipeline optimization.

In addition, I snuck in a change that adds the dual of the implicit
`write json` and `save file -` sinks to pipelines executed locally:
implicit `read json` and `load file -` sources. This makes it possible
to just write a transformation operator in most cases.