Introduce the `tenzir` and `tenzird` binaries
We're renaming VAST to Tenzir. This is the first PR of a series of PRs
that make this change. We want to make the change while causing as
little havoc as possible, and there is an announcement blog post coming
out soon that will explain what exactly is changing and why.

As a first step, this renames the `vast` binary to `tenzir`, and adds a
symlink `vast` that when used prints a warning on stderr about the
upcoming change.

As a second step, this adds `tenzird` and `tenzirctl` symlinks to the
`tenzir` binary.

- `tenzird` is a replacement for `vast start` and starts a node
- `tenzir` is a replacement for `vast exec` and executes a pipeline
- `tenzirctl` replaces all other commands until they're migrated to
  pipeline operators

This makes it clearer that Tenzir has two core functionalities: nodes
and pipelines.

It's an exciting time, and we at Tenzir are looking forward to making
this finally making this change.

Co-authored-by: Tobias Mayer <tobim@fastmail.fm>