Merge the namespace `vast::system` into `vast`
The namespace has historically made sense to keep a clear separation
between actor logic and application logic in the code. Nowadays, we
separate logic through plugins and builtins, and as such the question
what to put into which namespace became mostly obsolete. This removes
the nested namespace to avoid confusion.