Replace .markdownlint.yaml with .markdownlintrc
The reason is that markdownlint only looks recursively into parent
directories for a .markdownlintrc file. The YAML equivalent is only
looked at in the current directory. This is especially annoying when
writing documentation in an editor that integrates markdownlint, because
only when you launch the editor in the top-level repo, it will pick up
the correct configuration.