Revamp incoming request format in the node
A client forwarding incoming requests to the ndoe will generally
not be able to parse all request parameters to the correct
types declared by the endpoint. Therefore, for POST requests
we now forwarding the raw JSON body to the node, where it
can be parsed according to the correct endpoint description.

As a side effect the node became a bit more liberal in what it
accepts: Now everything that can be parsed as an instance of
the expected type is accepted. For example, `[1,2]` would be
accepted when the endpoints expects list of signed integers, a
list of unsigned integers or a list of strings. (but not e.g.
for a list of ip addresses)