Fix sporadic stalling of pipelines
This is the third iteration of pipeline execution. We've started with a
simple chaining of generators in a single thread, moved on to an
actor-based model that utilizes a push-based approach, and have since
come to realize a few downsides where we had simply made wrong
assumptions.

This change changes execution nodes to utilize a communication approach
closer to the pull-based Volcano model instead of CAF's push-based
streaming feature. This brings a few advantages:
1. Source operators are no longer polled in a busy loop. Tthis also
   applies to transformation or sink operators whose input is exhausted.
   This sometimes led to a stalling pipeline.
2. Back pressure now regulates the number events or bytes rather than
   number of variable-size batches of events or bytes transferred
   between operators.
3. The initialization order of operators is now well-defined: The
   execution nodes are spawned left-to-right, and the pipeline operators
   are guaranteed to be initialized right-to-left. Specifically, an
   operator must yield once before its preceding operator is
   initialized.
4. Operators can now await requests if they require a request to be
   responded to before continuing the pipeline. This works in a
   scheduler-friendly way.