Remove the enabled method from the plugin API.
A plugin being enabled or disabled based on the runtime config
is not useful except for component plugins.
Removing that bit of state from the generic plugin level rids us
form a couple of corner cases, e.g., missing plugins from the `version`
operator, or a warning that gets displayed `tenzir` connects to a
`tenzir-node` with mismatching plugins.