Add a library of support functions for functional tests
This most notably adds the `check` function that implements a
snapshot testing feature. Snapshot testing is a test style where
the expected outputs are stored separately from the test definitions
and can be regenerated by running the tests in an "update" mode.
The homegrown integration test framework that we are trying to replace
works in the same way.

The new library also adds some minor functions and a generic test
fixture that runs a single node.