Lay the foundation for pausing and resuming pipelines
This adds two new handlers to the pipeline executor for pausing and
resuming pipelines. This is yet untested, and has to be integrated into
the pipeline executor. However, the change itself is rather simple.

I wanted to get this out to unblock work on integrating this into the
pipeline manager API and into the app. Anyone should be able to take it
from here without needing to dive deep into the execution node's logic.