Work around broken dlopen for libfluent-bit.so
The fluent bit shared library contains a private copy of jemalloc
which can't be loaded with `dlopen()` at runtime. We work around
the issue by preloading fluent bit with the Tenzir binary when
possible.

The proper fix would be to use the system jemalloc instead of vendoring
a private copy in the fluent-bit package.