Fix program startup of the static binary
The recent addition of the s3 connector causes the linker to add
symbols from the AWS crypto abstraction layer library. This lib
uses the `dlopen` and `dlclose` functions even when built as a
static library - which is possible. What's more, the return value
of those functions is asserted, causing the tenzir executables to
abort during initialization.

The good news is that this can be worked around by disabling the
problematic assert. For that we pull a patch that does exactly that
from an open upstream pull request.