Temporarily use a custom package on debian
The official fluent-bit package contains a bundeld copy of jemalloc
that poses 2 problems: First, it is built in a way that makes it
impossible to load `libfluent-bit.so` with `dlopen()`. And second,
the way it initializes thread local variables creates problems with
the jemalloc present in `libarrow.so`, specifically it causes a
segfault when initializing the s3 filesystem.

I'm trying to resolve this issue upstream at
https://github.com/fluent/fluent-bit/pull/8005, but until it is
fixed we have to rely on this fallback.

Signed-off-by: Tobias Mayer <tobim@fastmail.fm>