Fix `serve` exiting prematurely
Pipelines ending with `serve` should wait until all events are fetched
from the corresponding `/serve` endpoint. This was not always the case
because of an incorrectly set scope in the operator.

Fixing that scope, however, led to the operator deadlocking the node
under some circumstances. This revealed a bug at the very core of CAF's
scheduling logic, which incorrectly blocked not just ordinary but also
system messages from arriving when a request response was awaited. I
got that bug fixed upstream and backported the fix to our bundled fork
of CAF 0.18. That in turn revealead us relying on the broken behavior in
our execution nodes accidentally. I've included a fix for that as well.