Add an optional content type to chunk
This addition allows for transferring metadata about chunks in-stream.
Specifically, this adds a metadata field for the content type, which is
the missing piece for tenzir/tenzir#3539.

The alternative to this approach is to forbid `write <format> | save
<connector>` in favor of always requiring the combined `to <connector>
write <format>`. However, the latter does not compose as freely as
transferring the content type in-stream.

What's left is to make use of the content type, but I will leave that to
the linked PR adding HTTP(S) savers.