Fix predicate pushdown for `export | where …`
This fixes a regression that snuck in between the v4.2 and v4.3
releases. We changed slightly how predicate pushdown works for the
`export` operator, which caused unnecessary predicates with trivially
true expressions to appear after predicate pushdownn. This exposed a bug
in the catalog, which did not bind expressions to the schemas before
evaluating them. The consequence of this was a too large set of
candidate partitions.

This pull request fixes both the unnecessary predicates in the
optimization phase of the `export` operator, and causes the catalog (1)
to normalize and validate expressions, and (2) to bind expressions
exactly once per schema, ignoring partitions for which tailoring fails.

Additionally, this properly sets the UUID for catalog lookup requests in
the `export` operator, which makes tracking the query's lifetime in
trace and debug logs easier.