Fix race in exec nodes that caused an infinite loop
This was a tricky one: With the recent changes to the behavior of await
in CAF that we backported, we had to switch to using `send` for the
internal run loop. That, however, unlike the direct scheduling on the
clock we used previously, can keep an actor alive. The scheduling of a
new run that this commit removes was unnecessary, and in fact actively
harmful as it led to an infinite loop during shutdown of an operator
when the timing was just right.