Remove flaky unit test
We haven't touched code that relates to this in ages, and it started to
fail sporadically in CI quite frequently over the past week. This
attempts to isolate or even fix the failure, that we are sure is some
form of UB.

I managed to reproduce the failure locally in a macOS VM that more or
less mirrors our CI environment.

After some investigation, @tobim and I agreed that it is impossible for
users to reach this code, as legacy types are only used as an
intermediate data structure during parsing, but not for any equality
comparisons outside of unit tests.